requirements.txt
# LESS视频标注工具依赖

# GUI框架
PyQt5>=5.15.0

# 视频处理
opencv-python>=4.5.0

# 数值计算
numpy>=1.21.0

# 图像处理 (用于中文文字渲染)
Pillow>=9.0.0

# 姿态检测 - MediaPipe
# Windows用户可能需要安装Visual C++ Redistributable
mediapipe>=0.10.0
